==> 01_MCQ_Processing/02.bott_processing.ipynb <==
import os
import numpy as np
import pandas as pd

file_list = [file for file in os.listdir("Data/Bott/") if file[-3:] == "txt" ]

file_list = [file for file in os.listdir("Data/Bott/") if file[-3:] == "txt" ]


split_lengths = [10, 32, 100, 32]

def split_string_with_length(string, split_length):
    splitted_string = []
    start = 0
    for length in split_length:
        splitted_string.append(string[start:start+length])
        start += length
    splitted_string.append(string[start:])   # take the rest of the string
    return splitted_string

dataframe_structure = {
    "File Name": [],
    "Serial #": [],
    "Litho Code 1": [],
    "MCQ Answers": [],
    "Litho Code 2": [],
    "Machine Generated #": []
}


for file in file_list:
    current_file = open("Data/Bott/"+file, "r")
    
    for line in current_file:
        if len(line) == 0:
            continue
        dataframe_structure["File Name"].append(file[:-4])
        
        line = line.replace(",", "")

        split_line = split_string_with_length(line[:len(line)-1], split_lengths)   # len(line)-1  - ignoring trailing new line character

        dataframe_structure["Serial #"].append(split_line[0])
        dataframe_structure["Litho Code 1"].append(split_line[1].replace(" ", "0"))
        dataframe_structure["MCQ Answers"].append(split_line[2])
        dataframe_structure["Litho Code 2"].append(split_line[3].replace(" ", "0"))
        dataframe_structure["Machine Generated #"].append(split_line[4])
        
    current_file.close()


df = pd.DataFrame(dataframe_structure)

# filter and seperate duplicated rows

Bott_Duplicate_Row = df[df.duplicated(['Litho Code 1', 'Litho Code 2'])]
Bott_Duplicate_Row.to_csv("Duplicate_Bott.csv", index=False, sep=';')

df = df.drop(Bott_Duplicate_Row.index)

# finalize bott processing

main_bott_dict = {
    "Serial_Bott": df["Serial #"],
    "File_Name_Bott": df["File Name"],
    "Answer": df["MCQ Answers"],
    "Litho_Bott_1": df["Litho Code 1"],
    "Litho_Bott_2": df["Litho Code 2"]
}

Main_Bott = pd.DataFrame(main_bott_dict)

Main_Bott.to_csv("Main_Bott.csv", index=False, sep=';')
