==> 01_MCQ_Processing/03.top_bott_processing.ipynb <==
import os
import numpy as np
import pandas as pd

main_top = pd.read_csv("Main_Top.csv", sep=';')
main_bott = pd.read_csv("Main_Bott.csv", sep=';')

top_litho_1 = main_top["Litho_Top_1"]
top_litho_2 = main_top["Litho_Top_2"]

bott_litho_1 = main_bott["Litho_Bott_1"]
bott_litho_2 = main_bott["Litho_Bott_2"]

top_litho_1_binary_equivalent_int = top_litho_1.copy()

for i in range(len(top_litho_1)):
    top_litho_1_binary_equivalent_int[i] = int(top_litho_1[i], 2)

top_litho_2 = top_litho_2[top_litho_1 != top_litho_2]

top_litho_2_binary_equivalent_int = top_litho_2.copy()

for i in range(len(top_litho_2)):
    top_litho_2_binary_equivalent_int.iloc[i] = int(top_litho_2.iloc[i], 2)

bott_litho_1_binary_equivalent_int = bott_litho_1.copy()

for i in range(len(bott_litho_1)):
    bott_litho_1_binary_equivalent_int[i] = int(bott_litho_1[i], 2)

bott_litho_2 = bott_litho_2[bott_litho_1 != bott_litho_2]

bott_litho_2_binary_equivalent_int = bott_litho_2.copy()

for i in range(len(bott_litho_2)):
    bott_litho_2_binary_equivalent_int.iloc[i] = int(bott_litho_2.iloc[i], 2)

top_litho_1_binary_equivalent_int = top_litho_1_binary_equivalent_int.sort_values()
top_litho_2_binary_equivalent_int = top_litho_2_binary_equivalent_int.sort_values()

bott_litho_1_binary_equivalent_int = bott_litho_1_binary_equivalent_int.sort_values()
bott_litho_2_binary_equivalent_int = bott_litho_2_binary_equivalent_int.sort_values()

# checking: litho_top_1 vs litho_bott_1

top_litho_int = top_litho_1_binary_equivalent_int
bott_litho_int = bott_litho_1_binary_equivalent_int


top_found_in_bott_index = []
bott_found_in_top_index = []

i, j = 0, 0

while True:
    
    if top_litho_int.iloc[i] == bott_litho_int.iloc[j]:
        top_found_in_bott_index.append(top_litho_int.index.values[i])
        bott_found_in_top_index.append(bott_litho_int.index.values[j])
        
        i += 1
        j += 1
    
    elif top_litho_int.iloc[i] < bott_litho_int.iloc[j]:
        i += 1
        
    else:
        j += 1
        
    if (i == len(top_litho_int)) or (j == len(bott_litho_int)):
        break

# keeping total match count

total_match = len(top_found_in_bott_index)

# dropping matched litho

top_litho_1_binary_equivalent_int = top_litho_1_binary_equivalent_int.drop(top_found_in_bott_index)
bott_litho_1_binary_equivalent_int = bott_litho_1_binary_equivalent_int.drop(bott_found_in_top_index)

# checking: top_litho_2 vs bott_litho_1

top_litho_int = top_litho_2_binary_equivalent_int
bott_litho_int = bott_litho_1_binary_equivalent_int


i, j = 0, 0

while True:
    
    if top_litho_int.iloc[i] == bott_litho_int.iloc[j]:
        top_found_in_bott_index.append(top_litho_int.index.values[i])
        bott_found_in_top_index.append(bott_litho_int.index.values[j])
        
        i += 1
        j += 1
    
    elif top_litho_int.iloc[i] < bott_litho_int.iloc[j]:
        i += 1
        
    else:
        j += 1
        
    if (i == len(top_litho_int)) or (j == len(bott_litho_int)):
        break

# count newly matched litho

newly_added = len(top_found_in_bott_index) - total_match
total_match = total_match + newly_added

# if anything newly matched then drop those

if newly_added != 0:    
    top_litho_2_binary_equivalent_int = top_litho_2_binary_equivalent_int.drop(top_found_in_bott_index[-newly_added:])
    bott_litho_1_binary_equivalent_int = bott_litho_1_binary_equivalent_int.drop(bott_found_in_top_index[-newly_added:])

# checking: litho_top_1 vs litho_bott_2

top_litho_int = top_litho_1_binary_equivalent_int
bott_litho_int = bott_litho_2_binary_equivalent_int


i, j = 0, 0

while True:
    
    if top_litho_int.iloc[i] == bott_litho_int.iloc[j]:
        top_found_in_bott_index.append(top_litho_int.index.values[i])
        bott_found_in_top_index.append(bott_litho_int.index.values[j])
        
        i += 1
        j += 1
    
    elif top_litho_int.iloc[i] < bott_litho_int.iloc[j]:
        i += 1
        
    else:
        j += 1
        
    if (i == len(top_litho_int)) or (j == len(bott_litho_int)):
        break

# count newly matched litho

newly_added = len(top_found_in_bott_index) - total_match
total_match = total_match + newly_added

# if anything newly matched then drop those

if newly_added != 0:    
    top_litho_1_binary_equivalent_int = top_litho_1_binary_equivalent_int.drop(top_found_in_bott_index[-newly_added:])
    bott_litho_2_binary_equivalent_int = bott_litho_2_binary_equivalent_int.drop(bott_found_in_top_index[-newly_added:])

# checking: top_litho_2 vs bott_litho_2

top_litho_int = top_litho_2_binary_equivalent_int
bott_litho_int = bott_litho_2_binary_equivalent_int


i, j = 0, 0

while True:
    
    if top_litho_int.iloc[i] == bott_litho_int.iloc[j]:
        top_found_in_bott_index.append(top_litho_int.index.values[i])
        bott_found_in_top_index.append(bott_litho_int.index.values[j])
        
        i += 1
        j += 1
    
    elif top_litho_int.iloc[i] < bott_litho_int.iloc[j]:
        i += 1
        
    else:
        j += 1
        
    if (i == len(top_litho_int)) or (j == len(bott_litho_int)):
        break

# count newly matched litho

newly_added = len(top_found_in_bott_index) - total_match
total_match = total_match + newly_added

# if anything newly matched then drop those

if newly_added != 0:
    top_litho_2_binary_equivalent_int = top_litho_2_binary_equivalent_int.drop(top_found_in_bott_index[-newly_added:])
    bott_litho_2_binary_equivalent_int = bott_litho_2_binary_equivalent_int.drop(bott_found_in_top_index[-newly_added:])

# selecting unmatched litho and giving output in files

main_top_not_found = main_top.drop(top_found_in_bott_index)
main_bott_not_found = main_bott.drop(bott_found_in_top_index)


main_top_not_found.to_csv("Top_Litho_Unmatchecd.csv", index=False, sep=';')
main_bott_not_found.to_csv("Bott_Litho_Unmatched.csv", index=False, sep=';')

# selecting matched litho, choosing file name and corresponding answers and giving output in file

main_top_found = main_top.loc[top_found_in_bott_index]
main_bott_found = main_bott.loc[bott_found_in_top_index]


bott_file_names = list(main_bott["File_Name_Bott"].loc[bott_found_in_top_index])
main_top_found["File_Name_Bott"] = bott_file_names


bott_answer = list(main_bott["Answer"].loc[bott_found_in_top_index])
main_top_found["Answer"] = bott_answer

# includeing main_bott litho

main_top_found["Litho_Bott_1"] = list(main_bott_found["Litho_Bott_1"])
main_top_found["Litho_Bott_2"] = list(main_bott_found["Litho_Bott_2"])


main_top_found.to_csv("Ready_for_Result_Processing.csv", index=False, sep=';')
