==> 01_MCQ_Processing/04.answer_processing.ipynb <==
import os
import numpy as np
import pandas as pd

split_lengths = [10, 1, 3, 100]

def split_string_with_length(string, split_length):
    splitted_string = []
    start = 0
    for length in split_length:
        splitted_string.append(string[start:start+length])
        start += length
    splitted_string.append(string[start:])   # take the rest of the string
    return splitted_string

dataframe_structure = {
    "Serial #": [],
    "Set_Code": [],
    "Subject_Code": [],
    "Answer": []
}


answer_file = open("Data/BOT_ANS.txt", "r")

for line in answer_file:
    answer_file_data = split_string_with_length(line, split_lengths)
    
    dataframe_structure["Serial #"].append(answer_file_data[0])
    dataframe_structure["Set_Code"].append(answer_file_data[1])
    dataframe_structure["Subject_Code"].append(answer_file_data[2])
    dataframe_structure["Answer"].append(answer_file_data[3])
    
answer_file.close()

answers = pd.DataFrame(dataframe_structure)

answer_200_1 = answers[(answers["Subject_Code"] == '200') & (answers["Set_Code"] == '1')]["Answer"].values[0]
answer_200_2 = answers[(answers["Subject_Code"] == '200') & (answers["Set_Code"] == '2')]["Answer"].values[0]
answer_200_3 = answers[(answers["Subject_Code"] == '200') & (answers["Set_Code"] == '3')]["Answer"].values[0]
answer_200_4 = answers[(answers["Subject_Code"] == '200') & (answers["Set_Code"] == '4')]["Answer"].values[0]

answer_300_1 = answers[(answers["Subject_Code"] == '300') & (answers["Set_Code"] == '1')]["Answer"].values[0]
answer_300_2 = answers[(answers["Subject_Code"] == '300') & (answers["Set_Code"] == '2')]["Answer"].values[0]
answer_300_3 = answers[(answers["Subject_Code"] == '300') & (answers["Set_Code"] == '3')]["Answer"].values[0]
answer_300_4 = answers[(answers["Subject_Code"] == '300') & (answers["Set_Code"] == '4')]["Answer"].values[0]

answer_400_1 = answers[(answers["Subject_Code"] == '400') & (answers["Set_Code"] == '1')]["Answer"].values[0]
answer_400_2 = answers[(answers["Subject_Code"] == '400') & (answers["Set_Code"] == '2')]["Answer"].values[0]
answer_400_3 = answers[(answers["Subject_Code"] == '400') & (answers["Set_Code"] == '3')]["Answer"].values[0]
answer_400_4 = answers[(answers["Subject_Code"] == '400') & (answers["Set_Code"] == '4')]["Answer"].values[0]

answer_list = [[], [], [0, answer_200_1, answer_200_2, answer_200_3, answer_200_4],
               [0, answer_300_1, answer_300_2, answer_300_3, answer_300_4],
               [0, answer_400_1, answer_400_2, answer_400_3, answer_400_4]]

def calculate_mark(answer, correct_answer):
    mark = 0
    
    for i in range(len(answer)):
        if correct_answer[i] == " ":
            mark += 1
        elif answer[i] == correct_answer[i]:
            mark += 1
        elif answer[i] == " ":
            mark += 0
        else:
            mark -= 0.5
            
    return mark

ready = pd.read_csv("Ready_for_Result_Processing.csv", sep=';')

student_subject_code = ready["Generated_Subject_Code"].astype(int)
student_set_code = ready["Set_Code"].astype(int)
student_answer = ready["Answer"].astype(str)

marks = []

for i in range(len(ready)):
    
    answer = student_answer[i]
    correct_answer = answer_list[int(student_subject_code[i]/100)][student_set_code[i]]
    
    mark = calculate_mark(answer, correct_answer)
    marks.append(mark)

ready["Marks"] = marks

df = {
    "Roll": ready["Roll"],
    "Marks": ready["Marks"],
    "Set_Code": ready["Set_Code"],
    "Subject_Code": ready["Subject_Code"],
    "Answer": ready["Answer"],
    "File_Name_Top": ready["File_Name_Top"],
    "Serial_Top": ready["Serial_Top"],
    "File_Name_Bott": ready["File_Name_Bott"],
    "Litho_Top_1": ready["Litho_Top_1"],
    "Litho_Top_2": ready["Litho_Top_1"],
    "Litho_Bott_1": ready["Litho_Bott_1"],
    "Litho_Bott_2": ready["Litho_Bott_2"]
}

tabulation_ready = pd.DataFrame(df)

tabulation_ready.to_csv("Tabulation.csv", index=False, sep=';')
